+++ packages/runtime-core/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./kiosk-app";
+export * from "./operational-control-controller";
 export * from "./transaction-controller";